Update all packets to support the new NetworkBuffer

+++ src/main/java/net/cg360/spookums/server/network/packet/type/PacketInOutEmpty.java
@@ -4,7 +4,7 @@ import net.cg360.spookums.server.network.packet.NetworkPacket;
 
 public abstract class PacketInOutEmpty extends NetworkPacket {
 
-    @Override protected final short encodeBody() { return 0; }
-    @Override protected final void decodeBody(short inboundSize) { }
+    @Override protected final int encodeBody() { return 0; }
+    @Override protected final void decodeBody(int inboundSize) { }
 
 }
